Move JCLI weighted pool IDs to separate module
