+++ jcli/src/jcli_app/certificate/mod.rs
@@ -10,6 +10,7 @@ mod get_stake_pool_id;
 mod new_stake_delegation;
 mod new_stake_pool_registration;
 mod sign;
+mod weighted_pool_ids;
 
 pub(crate) use self::sign::{pool_owner_sign, stake_delegation_account_binding_sign};
